Fix component name
