+++ app/components/HorizontalList/props.ts
@@ -2,8 +2,6 @@ export type dataObject = {
   title: string;
   subTitle: string;
   image?: string;
-  rating?: number;
-  description?: string;
 }
 
 export interface Props {
