+++ app/components/HorizontalList/style.ts
@@ -1,46 +1,27 @@
 import { StyleSheet } from 'react-native';
 import colors from '../../utils/colors';
-import width from '../../utils/commons';
 
 const styles = StyleSheet.create({
   container: {
-    width: width.deviceWidth,
+    width: 140,
     height: 180,
-    paddingLeft: 16,
-    paddingRight: 16,
-    flexDirection: 'row'
-  },
-  imageContainer: {
-    marginRight: 16,
+    marginLeft: 8,
+    marginRight: 8
   },
   image: {
     width: 140,
     height: 140,
-    resizeMode: 'contain'
-  },
-  contentContainer: {
-    flex: 1,
-    height: 140,
-    justifyContent: 'space-between'
+    justifyContent: 'flex-end',
+    padding: 5
   },
   subTitle: {
-    fontSize: 12,
-    color: colors.BLACK
+    fontSize: 14,
+    color: colors.WHITE
   },
   title: {
     fontSize: 14,
-    color: colors.BLACK,
-    fontWeight: '600'
-  },
-  button: {
-    backgroundColor: colors.RED,
-    borderColor: colors.RED
-  },
-  starContainer: {
-    width: 112
-  },
-  star: {
-    marginRight: 5
+    marginTop: 8,
+    color: colors.BLACK
   }
 });
