+++ app/components/VerticalList/index.tsx
@@ -1,20 +1,48 @@
 import React from 'react';
 import { Component } from 'react';
-import { FlatList, View, Text, ImageBackground } from 'react-native';
+import { FlatList, View, Text, Image, Platform } from 'react-native';
+import StarRating from 'react-native-star-rating';
 
 import { Props, dataObject } from './props';
 import { State } from './state';
 import styles from './style';
+import { MtrBtnNative, MtrBtnOpacity } from '../buttons';
+import colors from '../../utils/colors';
 
-class VerticalList extends Component<Props, State> {
+class HorizontalList extends Component<Props, State> {
+  renderButton = () => {
+    const title = 'VISIT';
+
+    if (Platform.OS === 'android') {
+      return <MtrBtnNative title={title} style={styles.button} />;
+    }
+
+    return <MtrBtnOpacity title={title} style={styles.button} />;
+  }
+
   renderItem = (data: any) => {
     const { item } = data;
     return (
       <View style={styles.container}>
-        <ImageBackground source={item.image} style={styles.image}>
+        <View style={styles.imageContainer}>
+          <Image source={item.image} style={styles.image} />
+        </View>
+        <View style={styles.contentContainer}>
+          <Text style={styles.title}>{item.title}</Text>
+          <StarRating
+            maxStars={5}
+            disabled
+            rating={item.rating}
+            fullStarColor={colors.RED}
+            halfStarEnabled
+            starSize={30}
+            containerStyle={styles.starContainer}
+            starStyle={styles.star}
+          />
           <Text style={styles.subTitle}>{item.subTitle}</Text>
-        </ImageBackground>
-        <Text style={styles.title}>{item.title}</Text>
+          <Text style={styles.subTitle}>{item.description}</Text>
+          {this.renderButton()}
+        </View>
       </View>
     );
   }
@@ -29,10 +57,9 @@ class VerticalList extends Component<Props, State> {
         data={data}
         keyExtractor={this.keyExtractor}
         renderItem={this.renderItem}
-        horizontal
       />
     );
   }
 }
 
-export default VerticalList;
+export default HorizontalList;
