+++ app/components/VerticalList/props.ts
@@ -2,6 +2,8 @@ export type dataObject = {
   title: string;
   subTitle: string;
   image?: string;
+  rating?: number;
+  description?: string;
 }
 
 export interface Props {
